Implementing Optimistic lock

+++ src/main/java/com/episen/ing3/tpmra/service/LockService.java
@@ -1,10 +1,11 @@
 package com.episen.ing3.tpmra.service;
 
 import java.time.OffsetDateTime;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import com.episen.ing3.tpmra.model.Lock;
@@ -19,39 +20,29 @@ public class LockService {
 	@Autowired
 	DocumentRepository documentRepository;
 
-	public Optional<Lock> getDocumentLock(Integer documentId) {
-		return lockRepository.findById(documentId);
+	public Lock getDocumentLock(Integer documentId) {
+		Optional<Lock> optionalLock = lockRepository.findById(documentId);
+		if(!optionalLock.isPresent()) throw new NoSuchElementException();
+		return optionalLock.get();
 
 	}
 
-	/*
-	 * TODO: treat the case were the lock already exists
-	 */
 	public Lock putDocumentLock(Integer documentId, String owner) {
+		if(lockRepository.findById(documentId).isPresent()) throw new AccessDeniedException("A lock is already put");
 		if(documentRepository.findById(documentId).isPresent()) {
 			OffsetDateTime dateTime = OffsetDateTime.now();
 			Lock lock = new Lock(documentId, owner, dateTime);
 			return lockRepository.save(lock);
 		}
-		return null;
+		else throw new NoSuchElementException();
 	}
 
-	public HttpStatus deleteDocumentLock(Integer documentId, String userName) {
+	public void deleteDocumentLock(Integer documentId, String userName) {
 		Optional<Lock> lock = lockRepository.findById(documentId);
 		if(lock.isPresent()) {
-			if(lock.get().getOwner().equals(userName)) {
-				lockRepository.deleteById(documentId);
-				if(getDocumentLock(documentId).isPresent())
-					return HttpStatus.INTERNAL_SERVER_ERROR;
-				else
-					return HttpStatus.NO_CONTENT;
-			}
-			else {
-				return HttpStatus.UNAUTHORIZED;
-			}
+			if(lock.get().getOwner().equals(userName)) lockRepository.deleteById(documentId);
+			else throw new AccessDeniedException("The lock was put by another user");
 		}
-		else {
-			return HttpStatus.NOT_FOUND;
-		}
+		else throw new NoSuchElementException();
 	}
 }
